feat: disallow usage of GPUs to CPU-only jobs (#96)
In Nomad jobs that didn't requested GPU, if the job landed in a GPU node and the job's docker image had CUDA libraries installed, then the job could access **all** the GPU of the node. Even if it didn't request any.

So for CPU-only jobs we set `NVIDIA_VISIBLE_DEVICES` to `None` to hide GPUs.

We also checked that jobs that asked one GPU got indeed only one GPU in the node (Nomad works fine in this case).